Remove /workbench/ dependency from test file in /editor/
